requirements.txt
python-telegram-bot[all]>=22.0,<23.0
motor==3.6.0
redis[asyncio]==5.1.1
pymongo==4.9.2
python-dotenv==1.1.1
httpx==0.28.1
